+++ app/ecosystem/page.tsx
@@ -40,10 +40,10 @@ const Ecosystem = async () => {
     }
   }
 
-  await updateAirtableRecords(recordsToUpdate)
+  // await updateAirtableRecords(recordsToUpdate)
 
   return (
-    <div className='pt-28'>
+    <div className='pt-28 min-h-[70vh]'>
       <div className='mb-[20px] md:mb-[60px]'>
         <Hero />
       </div>
